Simplify map projection

+++ app/src/test/java/com/kylecorry/trail_sense/tools/maps/domain/PhotoMapRotationServiceTest.kt
@@ -42,7 +42,7 @@ internal class PhotoMapRotationServiceTest {
     fun getCalibrationPoints(
         x: Float,
         y: Float,
-        rotation: Int,
+        rotation: Float,
         expectedX: Float,
         expectedY: Float
     ) {
